\name{ridgelines}
\alias{ridgelines}
\title{Draw Ridgeline Charts}
\description{
   Function to ridgeline charts.
}
\usage{
ridgelines(x, grp, xlim, ylim, xlab, adjust=1, n=TRUE, height=0.8, col, \dots)
}
\arguments{
   \item{x}{a numeric vector.}
   \item{grp}{the corresponding grouping variable.}
   \item{xlim}{...}
   \item{ylim}{...}
   \item{xlab}{...}
   \item{adjust}{...}
   \item{n}{...}
   \item{height}{...}
   \item{col}{...}
   \item{\dots}{other arguments.}
}
\details{
   The function draws a ridgeline chart for the numeric variable \code{x} for each level of the grouping variable \code{grp}.

   ...
}
\author{
   Wolfgang Viechtbauer \email{wvb@wvbauer.com} \url{https://www.wvbauer.com}
}
\examples{
   x <- 2
}
\keyword{hplot}
